python-service/requirements.txt
flask
flask_restful
twitch-python